requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
drf-yasg>=1.21.0
requests>=2.28.0
PyYAML>=6.0
uritemplate>=4.1.1
coreapi>=2.3.3
coreschema>=0.0.4
packaging>=21.0 